Straighten out assumptions and allow track switching.
I hadn't realised that the popup page is reloaded afresh each time and
only exists as long as it is visible. With that assumption now
straightened out things make much more sense. The status now reflects
the state of the audio when the popup loads accurately and will update
correctly based on the pausing/playing of the audio.